yolo_fourth_round_opt/requirements.txt
ultralytics>=8.4.14
pyyaml>=6.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
tqdm>=4.66.0
